+++ dIM/DataStore/Sources/DataStore/LiveDataController.swift
@@ -10,38 +10,6 @@ import CoreData
 import CoreBluetooth
 import Combine
 
-public enum DataControllerError: Error, LocalizedError {
-    case bluetoothTurnedOff
-    case sentEmptyMessage
-    case noConnectedDevices
-    case wrongDataFormat
-    case unknown
-    
-    public var errorDescription: String? {
-        switch self {
-        case .bluetoothTurnedOff:
-            return NSLocalizedString("You should turn Bluetooth on.", comment: "Bluetooth off")
-        case .sentEmptyMessage:
-            return NSLocalizedString("You cannot send empty messages.", comment: "No message text")
-        case .noConnectedDevices:
-            return NSLocalizedString("There are no connected devices.", comment: "No connection")
-        case .wrongDataFormat:
-            return NSLocalizedString("The message included invalid characters. Try another message.", comment: "JSON serialization failed")
-        default:
-            return NSLocalizedString("An unknown error has occured in the DataController.", comment: "Unknown error")
-        }
-    }
-}
-
-public protocol DataControllerDelegate: AnyObject {
-    func dataController(_ dataController: DataController, isConnectedTo deviceAmount: Int)
-    func dataController(_ dataController: DataController, didReceive encryptedMessage: Message)
-    func dataController(_ dataController: DataController, didReceiveAcknowledgement message: Message)
-    func dataController(_ dataController: DataController, didReceiveRead message: Message)
-    func dataController(_ dataController: DataController, didFailWith error: Error)
-    func dataControllerDidRelayMessage(_ dataController: DataController)
-}
-
 /// The responsibility of the DataController is to handle all Bluetooth related
 /// communication. All implementation details should be hidden from the outside,
 /// so much that we would not care if Bluetooth was used.
